DRY: make ColoringModel an abstract class with common methods and fields.
Only the feature mode part is specific.
