+++ src/main/java/org/mastodon/ui/coloring/ColoringModelBranchGraph.java
@@ -28,22 +28,15 @@
  */
 package org.mastodon.ui.coloring;
 
-import java.util.Optional;
-import java.util.stream.Stream;
-
 import org.mastodon.feature.FeatureModel;
 import org.mastodon.feature.FeatureProjection;
 import org.mastodon.graph.Edge;
 import org.mastodon.graph.Vertex;
 import org.mastodon.model.tag.TagSetModel;
-import org.mastodon.model.tag.TagSetStructure;
 import org.mastodon.ui.coloring.feature.FeatureColorMode;
 import org.mastodon.ui.coloring.feature.FeatureColorMode.EdgeColorMode;
 import org.mastodon.ui.coloring.feature.FeatureColorMode.VertexColorMode;
 import org.mastodon.ui.coloring.feature.FeatureColorModeManager;
-import org.mastodon.ui.coloring.feature.Projections;
-import org.mastodon.ui.coloring.feature.ProjectionsFromFeatureModel;
-import org.scijava.listeners.Listeners;
 
 /**
  * A ColoringModel that excludes modes defined for core-graph objects. The only
@@ -54,134 +47,18 @@ import org.scijava.listeners.Listeners;
  * @author Jean-Yves Tinevez
  */
 public class ColoringModelBranchGraph< V extends Vertex< E >, E extends Edge< V > >
-		implements TagSetModel.TagSetModelListener, FeatureColorModeManager.FeatureColorModesListener, ColoringModel
+		extends ColoringModel
+		implements TagSetModel.TagSetModelListener, FeatureColorModeManager.FeatureColorModesListener
 {
-
-	private final TagSetModel< ?, ? > tagSetModel;
-
-	private TagSetStructure.TagSet tagSet;
-
-	private FeatureColorMode featureColorMode;
-
-	private final FeatureColorModeManager featureColorModeManager;
-
-	private final Projections projections;
-
-	private final Listeners.List< ColoringChangedListener > listeners;
 
 	public ColoringModelBranchGraph(
 			final TagSetModel< ?, ? > tagSetModel,
 			final FeatureColorModeManager featureColorModeManager,
 			final FeatureModel featureModel )
 	{
-		this.tagSetModel = tagSetModel;
-		this.featureColorModeManager = featureColorModeManager;
-		this.projections = new ProjectionsFromFeatureModel( featureModel );
-		this.listeners = new Listeners.SynchronizedList<>();
+		super( tagSetModel, featureColorModeManager, featureModel );
 	}
 
-	@Override
-	public Listeners< ColoringChangedListener > listeners()
-	{
-		return listeners;
-	}
-
-	@Override
-	public void colorByNone()
-	{
-		tagSet = null;
-		featureColorMode = null;
-		listeners.list.forEach( ColoringChangedListener::coloringChanged );
-	}
-
-	@Override
-	public void colorByTagSet( final TagSetStructure.TagSet tagSet )
-	{
-		this.tagSet = tagSet;
-		this.featureColorMode = null;
-		listeners.list.forEach( ColoringChangedListener::coloringChanged );
-	}
-
-	@Override
-	public TagSetStructure.TagSet getTagSet()
-	{
-		return tagSet;
-	}
-
-	@Override
-	public void colorByFeature( final FeatureColorMode featureColorMode )
-	{
-		this.featureColorMode = featureColorMode;
-		this.tagSet = null;
-		listeners.list.forEach( ColoringChangedListener::coloringChanged );
-	}
-
-	@Override
-	public FeatureColorMode getFeatureColorMode()
-	{
-		return featureColorMode;
-	}
-
-	@Override
-	public boolean noColoring()
-	{
-		return tagSet == null && featureColorMode == null;
-	}
-
-	@Override
-	public void tagSetStructureChanged()
-	{
-		if ( tagSet != null )
-		{
-			final int id = tagSet.id();
-			final TagSetStructure tss = tagSetModel.getTagSetStructure();
-			final Optional< TagSetStructure.TagSet > ts = tss.getTagSets().stream().filter( t -> t.id() == id ).findFirst();
-			if ( ts.isPresent() )
-				colorByTagSet( ts.get() );
-			else
-				colorByNone();
-		}
-	}
-
-	@Override
-	public void featureColorModesChanged()
-	{
-		if ( featureColorMode != null )
-		{
-			final String name = featureColorMode.getName();
-			final Optional< FeatureColorMode > mode = Stream.concat(
-					featureColorModeManager.getBuiltinStyles().stream(),
-					featureColorModeManager.getUserStyles().stream() )
-					.filter( m -> m.getName().equals( name ) && isValid( m ) )
-					.findFirst();
-			if ( mode.isPresent() )
-				colorByFeature( mode.get() );
-			else
-				colorByNone();
-		}
-	}
-
-	@Override
-	public TagSetStructure getTagSetStructure()
-	{
-		return tagSetModel.getTagSetStructure();
-	}
-
-	@Override
-	public FeatureColorModeManager getFeatureColorModeManager()
-	{
-		return featureColorModeManager;
-	}
-
-	/**
-	 * Returns {@code true} if the specified color mode is valid against the
-	 * {@link FeatureModel}. That is: the feature projections that the color
-	 * mode rely on are declared in the feature model, and of the right class.
-	 *
-	 * @param mode
-	 *            the color mode
-	 * @return {@code true} if the color mode is valid.
-	 */
 	@Override
 	public boolean isValid( final FeatureColorMode mode )
 	{
