+++ src/main/java/org/mastodon/ui/coloring/ColoringModelMain.java
@@ -28,21 +28,14 @@
  */
 package org.mastodon.ui.coloring;
 
-import java.util.Optional;
-import java.util.stream.Stream;
-
 import org.mastodon.feature.FeatureModel;
 import org.mastodon.feature.FeatureProjection;
 import org.mastodon.graph.Edge;
 import org.mastodon.graph.Vertex;
 import org.mastodon.graph.branch.BranchGraph;
 import org.mastodon.model.tag.TagSetModel;
-import org.mastodon.model.tag.TagSetStructure;
 import org.mastodon.ui.coloring.feature.FeatureColorMode;
 import org.mastodon.ui.coloring.feature.FeatureColorModeManager;
-import org.mastodon.ui.coloring.feature.Projections;
-import org.mastodon.ui.coloring.feature.ProjectionsFromFeatureModel;
-import org.scijava.listeners.Listeners;
 
 /**
  * ColoringModel knows which coloring scheme is currently active. Possible
@@ -64,20 +57,9 @@ public class ColoringModelMain<
 	E extends Edge<V>,
 	BV extends Vertex<BE>,
 	BE extends Edge< BV > >
-		implements TagSetModel.TagSetModelListener, FeatureColorModeManager.FeatureColorModesListener, ColoringModel
+		extends ColoringModel
+		implements TagSetModel.TagSetModelListener, FeatureColorModeManager.FeatureColorModesListener
 {
-
-	private final TagSetModel< ?, ? > tagSetModel;
-
-	private TagSetStructure.TagSet tagSet;
-
-	private FeatureColorMode featureColorMode;
-
-	private final FeatureColorModeManager featureColorModeManager;
-
-	private final Projections projections;
-
-	private final Listeners.List< ColoringChangedListener > listeners;
 
 	private final BranchGraph< BV, BE, V, E > branchGraph;
 
@@ -87,115 +69,10 @@ public class ColoringModelMain<
 			final FeatureModel featureModel,
 			final BranchGraph< BV, BE, V, E > branchGraph )
 	{
-		this.tagSetModel = tagSetModel;
-		this.featureColorModeManager = featureColorModeManager;
+		super( tagSetModel, featureColorModeManager, featureModel );
 		this.branchGraph = branchGraph;
-		this.projections = new ProjectionsFromFeatureModel( featureModel );
-		this.listeners = new Listeners.SynchronizedList<>();
 	}
 
-	@Override
-	public Listeners< ColoringChangedListener > listeners()
-	{
-		return listeners;
-	}
-
-	@Override
-	public void colorByNone()
-	{
-		tagSet = null;
-		featureColorMode = null;
-		listeners.list.forEach( ColoringChangedListener::coloringChanged );
-	}
-
-	@Override
-	public void colorByTagSet( final TagSetStructure.TagSet tagSet )
-	{
-		this.tagSet = tagSet;
-		this.featureColorMode = null;
-		listeners.list.forEach( ColoringChangedListener::coloringChanged );
-	}
-
-	@Override
-	public TagSetStructure.TagSet getTagSet()
-	{
-		return tagSet;
-	}
-
-	@Override
-	public void colorByFeature( final FeatureColorMode featureColorMode )
-	{
-		this.featureColorMode = featureColorMode;
-		this.tagSet = null;
-		listeners.list.forEach( ColoringChangedListener::coloringChanged );
-	}
-
-	@Override
-	public FeatureColorMode getFeatureColorMode()
-	{
-		return featureColorMode;
-	}
-
-	@Override
-	public boolean noColoring()
-	{
-		return tagSet == null && featureColorMode == null;
-	}
-
-	@Override
-	public void tagSetStructureChanged()
-	{
-		if ( tagSet != null )
-		{
-			final int id = tagSet.id();
-			final TagSetStructure tss = tagSetModel.getTagSetStructure();
-			final Optional< TagSetStructure.TagSet > ts = tss.getTagSets().stream().filter( t -> t.id() == id ).findFirst();
-			if ( ts.isPresent() )
-				colorByTagSet( ts.get() );
-			else
-				colorByNone();
-		}
-	}
-
-	@Override
-	public void featureColorModesChanged()
-	{
-		if ( featureColorMode != null )
-		{
-			final String name = featureColorMode.getName();
-			final Optional< FeatureColorMode > mode = Stream.concat(
-					featureColorModeManager.getBuiltinStyles().stream(),
-					featureColorModeManager.getUserStyles().stream() )
-					.filter( m -> m.getName().equals( name ) && isValid( m ) )
-					.findFirst();
-			if ( mode.isPresent() )
-				colorByFeature( mode.get() );
-			else
-				colorByNone();
-		}
-	}
-
-	@Override
-	public TagSetStructure getTagSetStructure()
-	{
-		return tagSetModel.getTagSetStructure();
-	}
-
-	@Override
-	public FeatureColorModeManager getFeatureColorModeManager()
-	{
-		return featureColorModeManager;
-	}
-
-	/**
-	 * Returns {@code true} if the specified color mode is valid against the
-	 * {@link FeatureModel}. That is: the feature projections that the color
-	 * mode rely on are declared in the feature model, and of the right class.
-	 *
-	 * @param mode
-	 *            the color mode
-	 * @return {@code true} if the color mode is valid.
-	 */
 	@Override
 	public boolean isValid( final FeatureColorMode mode )
 	{
